Some but not all unit tests refactored; webook validation and authentication needs sorting out.
