+++ pusher/errors.py
@@ -1,5 +1,12 @@
 # -*- coding: utf-8 -*-
 
+from __future__ import (
+    print_function,
+    unicode_literals,
+    absolute_import,
+    division)
+
+
 class PusherError(Exception):
     pass
