+++ pusher/signature.py
@@ -1,5 +1,11 @@
 # -*- coding: utf-8 -*-
 
+from __future__ import (
+    print_function,
+    unicode_literals,
+    absolute_import,
+    division)
+
 import hashlib
 import hmac
 import six
